refactor(notebook): Centralize cell action handling and simplify action bar
Refactors action handling to be more consistent across all action positions.
Removes the needsCellContext flag as all actions now ensure proper cell
selection before execution. Simplifies the action bar component by using
the new useActionsForCell hook.

Changes:
- Refactor NotebookCellActionBar to use useActionsForCell hook
- Update CellActionButton to handle cell selection uniformly
- Simplify NotebookCellMoreActionsMenu to always select cell before action
- Remove needsCellContext checks in actionBarMenuItems
- Use centralized selection status from cell observables

Part of #9262
