+++ src/vs/workbench/contrib/positronNotebook/browser/notebookCells/NotebookCellActionBar.tsx
@@ -11,65 +11,30 @@ import React, { useState } from 'react';
 
 // Other dependencies.
 import { localize } from '../../../../../nls.js';
-import { usePositronReactServicesContext } from '../../../../../base/browser/positronReactRendererContext.js';
-import { IPositronNotebookCell } from '../PositronNotebookCells/IPositronNotebookCell.js';
-import { ActionButton } from '../utilityComponents/ActionButton.js';
+import { CellSelectionStatus, IPositronNotebookCell } from '../PositronNotebookCells/IPositronNotebookCell.js';
 import { useNotebookInstance } from '../NotebookInstanceProvider.js';
-import { useSelectionStatus } from './useSelectionStatus.js';
 import { NotebookCellMoreActionsMenu } from './actionBar/NotebookCellMoreActionsMenu.js';
-import { useActionBarVisibility } from './actionBar/useActionBarVisibility.js';
-import { NotebookCellActionBarRegistry, INotebookCellActionBarItem } from './actionBar/actionBarRegistry.js';
+import { useActionsForCell } from './actionBar/useActionsForCell.js';
+import { CellActionButton } from './actionBar/CellActionButton.js';
 import { useObservedValue } from '../useObservedValue.js';
-import { CellSelectionType } from '../../../../services/positronNotebook/browser/selectionMachine.js';
-import { createCellInfo } from './actionBar/cellConditions.js';
-
 
 interface NotebookCellActionBarProps {
 	cell: IPositronNotebookCell;
 	children: React.ReactNode;
-	isHovered: boolean;
 }
 
-export function NotebookCellActionBar({ cell, children, isHovered }: NotebookCellActionBarProps) {
-	const services = usePositronReactServicesContext();
-	const commandService = services.commandService;
+export function NotebookCellActionBar({ cell, children }: NotebookCellActionBarProps) {
+	const actionsForCell = useActionsForCell(cell);
 	const instance = useNotebookInstance();
-	const registry = NotebookCellActionBarRegistry.getInstance();
+	const mainActions = actionsForCell.main;
+	const menuActions = actionsForCell.menu;
 	const [isMenuOpen, setIsMenuOpen] = useState(false);
-	const selectionStatus = useSelectionStatus(cell);
-
-	// Use observable values for reactive updates
-	const allMainActions = useObservedValue(registry.mainActions) ?? [];
-	const allMenuActions = useObservedValue(registry.menuActions) ?? [];
-
-	// Filter actions based on cell conditions
-	const cells = instance.cells.get();
-	const cellIndex = cells.indexOf(cell);
-	const cellInfo = createCellInfo(cell, cellIndex, cells.length);
-
-	const mainActions = allMainActions.filter(action => {
-		return !action.cellCondition || action.cellCondition(cellInfo);
-	});
-
-	const menuActions = allMenuActions.filter(action => {
-		return !action.cellCondition || action.cellCondition(cellInfo);
-	});
+	const selectionStatus = useObservedValue(cell.selectionStatus);
 
 	const hasMenuActions = menuActions.length > 0;
 
 	// Determine visibility using the extracted hook
-	const shouldShowActionBar = useActionBarVisibility(isHovered, isMenuOpen, selectionStatus);
-
-	const handleActionClick = (action: INotebookCellActionBarItem) => {
-		// If action needs cell context, ensure cell is selected first
-		if (action.needsCellContext) {
-			// Select the cell (the method handles checking if already selected)
-			instance.selectionStateMachine.selectCell(cell, CellSelectionType.Normal);
-		}
-
-		// Execute the command (without passing cell as argument)
-		commandService.executeCommand(action.commandId);
-	};
+	const shouldShowActionBar = isMenuOpen || selectionStatus === CellSelectionStatus.Selected || selectionStatus === CellSelectionStatus.Editing;
 
 	return <div
 		aria-hidden={!shouldShowActionBar}
@@ -82,20 +47,17 @@ export function NotebookCellActionBar({ cell, children, isHovered }: NotebookCel
 
 		{/* Render contributed main actions - will auto-update when registry changes */}
 		{mainActions.map(action => (
-			<ActionButton
+			<CellActionButton
 				key={action.commandId}
-				ariaLabel={String(action.label ?? action.commandId)}
-				onPressed={() => handleActionClick(action)}
-			>
-				<div className={`button-icon codicon ${action.icon}`} />
-			</ActionButton>
+				action={action}
+				cell={cell}
+			/>
 		))}
 
 		{/* Dropdown menu for additional actions - only render if there are menu actions */}
 		{hasMenuActions ? (
 			<NotebookCellMoreActionsMenu
 				cell={cell}
-				commandService={commandService}
 				instance={instance}
 				menuActions={menuActions}
 				onMenuStateChange={setIsMenuOpen}
